Убрал везде int, потому что не влезает тот же id чата

+++ New/TelegramWorkLibrary/TelegramWorkLibrary/Result.cs
@@ -8,7 +8,7 @@ namespace TelegramWorkLibrary
 {
     public class Result
     {
-        public int _updateId { get; set; } 
+        public Int64 _updateId { get; set; } 
         public Message _message { get; set; } // Сообщение
     }
 }
