+++ New/TelegramWorkLibrary/TelegramWorkLibrary/Struct/Message.cs
@@ -8,8 +8,8 @@ namespace TelegramWorkLibrary
 {
     public class Message
     {
-        public int _messageId { get; set; } // Идентификатор сообщения
-        public int _date { get; set; } // Дата отправки сообщения (unix time)
+        public Int64 _messageId { get; set; } // Идентификатор сообщения
+        public Int64 _date { get; set; } // Дата отправки сообщения (unix time)
         public String _text { get; set; } // Текст сообщения
 
         public User _from {get; set;} // Отправитель (Может быть пустым)
@@ -34,11 +34,11 @@ namespace TelegramWorkLibrary
         public bool? _groupChatCreated { get; set; } // Опционально. Сервисное сообщение: группа создана 
         public bool? _supergroupChatCreated { get; set; } // Опционально. Сервисное сообщение: супергруппа создана 
         public bool? _channelChatCreated { get; set; } // Опционально. Сервисное сообщение: канал создан
-        public int? _migrateToChatId { get; set; } // Опционально. Группа была преобразована в супергруппу с указанным идентификатором. Не превышает 1e13 
-        public int? _migrateFromChatId { get; set; } // Опционально. Cупергруппа была создана из группы с указанным идентификатором. Не превышает 1e13 
+        public Int64? _migrateToChatId { get; set; } // Опционально. Группа была преобразована в супергруппу с указанным идентификатором. Не превышает 1e13 
+        public Int64? _migrateFromChatId { get; set; } // Опционально. Cупергруппа была создана из группы с указанным идентификатором. Не превышает 1e13 
 
         public User _forwardFrom { get; set; } // Опционально. Для пересланных сообщений: отправитель оригинального сообщения 
-        public int? _forwardDate { get; set; } // Опционально. Для пересланных сообщений: дата отправки оригинального сообщения
+        public Int64? _forwardDate { get; set; } // Опционально. Для пересланных сообщений: дата отправки оригинального сообщения
         public Message _replyToMessage { get; set; } // Опционально. Для ответов: оригинальное сообщение.
         public Message _pinnedMessage { get; set; } // Опционально. Указанное сообщение было прикреплено.
