+++ New/TelegramWorkLibrary/TelegramWorkLibrary/Struct/MessageEntity.cs
@@ -10,8 +10,8 @@ namespace TelegramWorkLibrary
     public class MessageEntity
     {
         public TypeEntity _type { get; set; } // Тип сущности 
-        public int _offset { get; set; } // Смещение элементов кода utf-16 в начале сущности
-        public int _length { get; set; } // Длина объекта в utf-16 единиц код 
+        public Int64 _offset { get; set; } // Смещение элементов кода utf-16 в начале сущности
+        public Int64 _length { get; set; } // Длина объекта в utf-16 единиц код 
         public String _url { get; set; } // Опционально. Для “text_link” только URL-адрес, который будет открыт после того, как пользователь нажимает на текст
     }
 }
