+++ New/TelegramWorkLibrary/TelegramWorkLibrary/Struct/User.cs
@@ -8,7 +8,7 @@ namespace TelegramWorkLibrary
 {
     public class User
     {
-        public int _id { get; set; } // идентификатор бота 
+        public Int64 _id { get; set; } // идентификатор бота 
         public String _firstName { get; set; } // Имя пользователя 
         public String _lastName { get; set; } // Фамилия пользователя
         public String _userName { get; set; } // Логин пользователя
